requirements.txt
colorama
crypto
fade
mysql-connector-python
nltk
password-strength
progressbar
pyperclip
pwnedpasswords
random-word
rich
emoji
pycryptodome
getpass
